+++ gauss_filter.h
@@ -1,8 +1,8 @@
-#include <stdio.h>
 #include <math.h>
+#include <stdio.h>
 
 #define kernel_size 1000
 #define sigma 1.2
 #define PI 3.1415926535
 
-double **make_gauss_kernel();
+int **make_gauss_kernel();
